pyproject.toml
[project]
name = "obsidian-llm-tool-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.21",
    "dspy-ai>=2.6.21",
    "ipykernel>=6.29.5",
    "nbconvert>=7.16.6",
    "openai>=1.76.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "youtube-transcript-api>=1.1.0",
]
